+++ src/app/layout.tsx
@@ -1,23 +1,13 @@
-import type { Metadata } from 'next'
-import { AppShell } from '@/components/layout/AppShell'
 import { Providers } from './providers'
+import Main from './main'
 import './globals.css'
 
-export const metadata: Metadata = {
-  title: 'Deep 自販機 Map',
-  description: '自動販売機マップ',
-}
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode
-}>) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="ja">
-      <body>
+      <body style={{ display: 'flex', flexDirection: 'column', minHeight: '100vh' }}>
         <Providers>
-          <AppShell>{children}</AppShell>
+          <Main>{children}</Main>
         </Providers>
       </body>
     </html>
