+++ src/app/providers.tsx
@@ -1,25 +1,15 @@
 'use client'
 
 import { ChakraProvider } from '@chakra-ui/react'
-import { LoadScript } from '@react-google-maps/api'
 import { UserProvider } from '@/Contexts/UserContext'
-import { GOOGLE_MAPS_API_KEY, GOOGLE_MAPS_MAP_ID } from '@/config/env'
-import { GOOGLE_MAP_LIBRARIES } from '@/config/googleMaps'
+import { SearchProvider } from '@/Contexts/SearchContext'
 
 export function Providers({ children }: { children: React.ReactNode }) {
-  if (!GOOGLE_MAPS_API_KEY) {
-    throw new Error('NEXT_PUBLIC_GOOGLE_MAPS_API_KEY ERROR')
-  }
-
   return (
-    <LoadScript
-      googleMapsApiKey={GOOGLE_MAPS_API_KEY}
-      libraries={GOOGLE_MAP_LIBRARIES}
-      mapIds={[GOOGLE_MAPS_MAP_ID]}
-    >
-      <UserProvider>
+    <UserProvider>
+      <SearchProvider>
         <ChakraProvider>{children}</ChakraProvider>
-      </UserProvider>
-    </LoadScript>
+      </SearchProvider>
+    </UserProvider>
   )
 }
